refactor: import, export of constants
