+++ src/components/AddressModal/AddressModal.jsx
@@ -1,6 +1,6 @@
 import { v4 as uuid } from "uuid";
 import { useEffect } from "react";
-import addressActionTypes from "../../constants/addressActionTypes";
+import {addressActionTypes} from "../../constants/constants";
 import { useAddress } from "../../contexts/address-context";
 import "./AddressModal.css";
