--- src/constants/addressActionTypes.js
@@ -1,13 +1,0 @@
-const addressActionTypes = {
-    DISPLAY_ADDRESSES: "DISPLAY_ADDRESSES",
-    SHOW_ADDRESS_MODAL: "SHOW_ADDRESS_MODAL",
-    SET_ADDRESS_DETAILS: "SET_ADDRESS_DETAILS",
-    EDIT_ADDRESS_DETAILS: "EDIT_ADDRESS_DETAILS",
-    ADD_NEW_ADDRESS: "ADD_NEW_ADDRESS",
-    SET_DUMMY_ADDRESS: "SET_DUMMY_ADDRESS",
-    REMOVE_ADDRESS: "REMOVE_ADDRESS",
-    RESET_ADDRESS_FORM: "RESET_ADDRESS_FORM",
-    SET_SELECTED_ADDRESS_ID: "SET_SELECTED_ADDRESS_ID",
-}
-
-export default addressActionTypes;
