--- src/constants/cartActionTypes.js
@@ -1,8 +1,0 @@
-const cartActionTypes = {
-    DISPLAY_CART: "DISPLAY_CART",
-    ADD_TO_CART: "ADD_TO_CART",
-    REMOVE_FROM_CART: "REMOVE_FROM_CART",
-    UPDATE_QUANTITY_IN_CART: "UPDATE_QUANTITY_IN_CART",
-}
-
-export default cartActionTypes;
