--- src/constants/filterActionTypes.js
@@ -1,13 +1,0 @@
-const filterActionTypes = {
-  DISPLAY_PRODUCTS: "DISPLAY_PRODUCTS",
-  DISPLAY_CATEGORIES: "DISPLAY_CATEGORIES",
-  SEARCH: "SEARCH",
-  CATEGORY: "CATEGORY",
-  SIZE: "SIZE",
-  AVAILABILITY: "AVAILABILITY",
-  SORT_BY_PRICE: "SORT_BY_PRICE",
-  SORT_BY_RATING_RANGE: "SORT_BY_RATING_RANGE",
-  CLEAR_FILTERS: "CLEAR_FILTERS",
-};
-
-export default filterActionTypes;
