--- src/constants/wishlistActionTypes.js
@@ -1,7 +1,0 @@
-const wishlistActionTypes ={
-    DISPLAY_WISHLIST: "DISPLAY_WISHLIST",
-    ADD_TO_WISHLIST: "ADD_TO_WISHLIST",
-    REMOVE_FROM_WISHLIST: "REMOVE_FROM_WISHLIST",
-}
-
-export default wishlistActionTypes;
